requirements.txt
# AI Image Generator - Dependencies
# Python 3.9+ required

# Qt GUI Framework
PySide6>=6.5.0

# HTTP Client for API calls
requests>=2.31.0

# OpenAI API (ChatGPT)
openai>=1.0.0

# Google Generative AI (Gemini)
google-generativeai>=0.3.0

# Image processing
Pillow>=10.0.0

# Async support
aiohttp>=3.9.0
